src/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
supabase==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-multipart==0.0.9
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.1
pydantic==2.6.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==8.0.1
pytest-asyncio==0.23.5
gunicorn==21.2.0
beautifulsoup4==4.12.3
huggingface_hub==0.20.3
mailchimp-marketing==3.0.80
convertkit==0.2.0
sentry-sdk==1.40.0
stripe==7.12.0
python-twitter==3.5.0
openai==1.12.0
anthropic==0.18.1
replicate==0.24.0
deepgram-sdk==3.1.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis[hiredis]==5.0.1
flower==2.0.1
postgrest==0.13.1
gotrue==1.3.0
realtime==1.0.0 